all: disable any -race test that fails or times out
This PR disables any package that fails under the race detector (ie, has a race)
or times out due to internal errors.
