+++ cmd/jujud/agent/machine_test.go
@@ -12,7 +12,6 @@ import (
 	"runtime"
 	"strings"
 	"sync/atomic"
-	"testing"
 	"time"
 
 	"github.com/juju/cmd"
@@ -94,12 +93,6 @@ var (
 	_ = gc.Suite(&MachineWithCharmsSuite{})
 	_ = gc.Suite(&mongoSuite{})
 )
-
-func TestPackage(t *testing.T) {
-	// TODO(waigani) 2014-03-19 bug 1294458
-	// Refactor to use base suites
-	coretesting.MgoTestPackage(t)
-}
 
 type commonMachineSuite struct {
 	singularRecord *singularRunnerRecord
