+++ worker/provisioner/package_test.go
@@ -6,9 +6,14 @@ package provisioner_test
 import (
 	stdtesting "testing"
 
-	"github.com/juju/juju/testing"
+	"github.com/juju/testing"
+
+	coretesting "github.com/juju/juju/testing"
 )
 
-func Test(t *stdtesting.T) {
-	testing.MgoTestPackage(t)
+func TestPackage(t *stdtesting.T) {
+	if testing.RaceEnabled {
+		t.Skip("skipping package under -race, see LP 1519097")
+	}
+	coretesting.MgoTestPackage(t)
 }
